Guard functions in fmpfr_glue.c; sort interfaces alphabetically.
This fixes a bug in the generating script where code in fmpfr_glue.c
was not guarded by the appropriate #if USE_ ... statements.

I have also put some "sort" commands into the generate script
so that interfaces are not randomly re-sorted each invocation.
For this patch, that doesn't help, but in the future it should
be better.